Updated Game_Music_Emu

+++ plugins/gme/game-music-emu-0.6pre/gme/Ym2413_Emu.h
@@ -1,6 +1,6 @@
 // YM2413 FM sound chip emulator interface
 
-// Game_Music_Emu 0.6-pre
+// Game_Music_Emu $vers
 #ifndef YM2413_EMU_H
 #define YM2413_EMU_H
 
@@ -16,7 +16,7 @@ public:
 	
 	// Sets output sample rate and chip clock rates, in Hz. Returns non-zero
 	// if error.
-	int set_rate( double sample_rate, double clock_rate );
+	int set_rate( int sample_rate, int clock_rate );
 	
 	// Resets to power-up state
 	void reset();
